Changes to uefi decompress functions
Created hal.uefi_fv to contain uefi_fv functions
Modified Guids to use UUID
function cleanup
